+++ src/components/schedule/delete.php
@@ -21,11 +21,12 @@ if(!isset(\CMSApplication::$VAR['schedule_id']) || !\CMSApplication::$VAR['sched
 }
 
 // Get workorder_id before deleting the record
-\CMSApplication::$VAR['workorder_id'] = $this->app->components->schedule->getRecord(\CMSApplication::$VAR['schedule_id'], 'workorder_id');
+// \CMSApplication::$VAR['workorder_id'] = $this->app->components->schedule->getRecord(\CMSApplication::$VAR['schedule_id'], 'workorder_id');
 
-// Delete the schedule
-$this->app->components->schedule->deleteRecord(\CMSApplication::$VAR['schedule_id']);
-
-// load schedule search page
-$this->app->system->variables->systemMessagesWrite('success', _gettext("Schedule record has been deleted."));
-$this->app->system->page->forcePage('workorder', 'details&workorder_id='.\CMSApplication::$VAR['workorder_id']);
+// Run the delete function if allowed
+if(!$this->app->components->schedule->checkRecordAllowsDelete(\CMSApplication::$VAR['schedule_id'])) {
+    $this->app->system->page->forcePage('schedule', 'details&schedule_id='.\CMSApplication::$VAR['schedule_id']);
+} else {
+    $this->app->components->schedule->deleteRecord(\CMSApplication::$VAR['schedule_id']);
+    $this->app->system->page->forcePage('schedule', 'search');
+}
